Rename pulp.platform to pulp.app
closes #2226
https://pulp.plan.io/issues/2226
