+++ app/pulp/app/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'pulp.app.apps.PulpAppConfig'
