+++ app/pulp/app/tests/test_settings.py
@@ -1,6 +1,6 @@
 from unittest import TestCase, mock
 
-from pulp.platform import settings as pulp_settings
+from pulp.app import settings as pulp_settings
 
 
 class TestMergeSettings(TestCase):
@@ -55,7 +55,7 @@ class TestLoadSettings(TestCase):
         expected['allowed_hosts'] = ['subdomains.all.the.way.down.www.example.com']
 
         mocked_open = mock.mock_open(read_data=override)
-        with mock.patch('pulp.platform.settings.open', mocked_open, create=True):
+        with mock.patch('pulp.app.settings.open', mocked_open, create=True):
             settings = pulp_settings.load_settings('somefile')
 
         self.assertEqual(settings, expected)
