--- platform/pulp/platform/apps.py
@@ -1,6 +1,0 @@
-from django import apps
-
-
-class PulpPlatformConfig(apps.AppConfig):
-    name = 'pulp.platform'
-    label = 'pulp_platform'
